+++ src/WorkerStatus.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace ShopWatch\MessengerWorkerRegistry;
+
+enum WorkerStatus: string
+{
+    case Running = 'running';
+    case Stopped = 'stopped';
+    case Dead = 'dead';
+}